Push object format improvements
Two big changes:

1. The clientId is no longer included in the push object.

This removes a big implementation difficulty for servers (especially big
distributed mail stores), and allows the server to easily coalesce/rate-limit
push events.

2. The current state strings for each object types are now included.

This is rather than just a single state string representing
mail/contacts/calendars. This allows the client to more efficiently call
getFooUpdates just for those types that need it, and to avoid the need to make
any calls if the push is due to the client's own setFoos call.

Since the state strings are the same as returned by the getFoo calls, this also
removes a race condition on startup, where you either had to wait for the push
connection to establish first before requesting any data, or you had to
immediately call getFooUpdates when the push connection establishes, since there
was no way of knowing if the "mail" state (for example) corresponded with your
loaded mailboxes/threads/messages state.