Move JMAP version/extensions from Account obj to auth result.
The client will need a version or extension in order to be able to fetch future
versions of the Account object! It's unlikely the available option will vary
between accounts (and it can still vary between logged in users, as it is only
returned along with the URLs at the end of the authentication process).