Clients must empty a mailbox before deleting it
This fixes one area of the spec where the client could still lock up the server
with a potentially huge expensive change. It was also unique in that it would affect another data type implicitly, whereas all changes in JMAP are normally explicit. Forbidding non-empty mailboxes from being destroyed fixes this.

(A client can of course still allow a non-empty mailbox to be destroyed, it just has to remove the messages explicitly as well before destroying the mailbox via JMAP.)