Remove the Message#inReplyToMessageId property
This property is problematic, as it is meant to be immutable but it
references a JMAP message id, and the message may arrive or be deleted
after this property is first set.

Instead, just use the headers directly in the spec to specify how to
set the isAnswered property after sending replies and for sorting
messages in a thread.