Allow update to patch sub-parts of object.
The update model has always been a patch, but before you could only omit
properties at the top-level. This change allows properties to be changed
on sub-objects.

This allows you to add a keyword without caring what other keywords are
currently on the object. Or to update just one property on a deeply
nested object structure without sending a whole lot of unnecessary data
(this will be particularly relevant when extended to calendars, where
event objects have a much deeper object structure).

Resolves #152