Replace Message#date with receivedAt/sentAt
The Message#date property corresponded to the IMAP internal date. The spec now
makes this explicit (as the recievedAt property) and allows the client to
access a parsed from of the  message date header as well (as the sentAt
property).

Most dates in JMAP are UTC, but the date header may include time zone offset
information. The Date datatype definition has been split into UTCDate and Date
to allow this to be returned to the client, while restricting other values that
should remain UTC only.

Resolves #118