Remove width/height properties from EmailBodyPart
These were already optional, since they require a lot of extra work on the
server, and are at odds with the rest of the properties which are all simply
about the MIME structure. It seems these are better as optional/custom
extensions to JMAP rather than part of the core spec.