Include images not in both body lists in attachedFiles
Some emails with poor MIME structure have images that would appear in (for
example) the htmlBody list of parts, but not textBody. These should be included
in the attachedFiles list; clients can easily skip them if they are in the body
list it is using, and otherwise it may miss displaying important data.