Rewrite primaryAccounts definition for clarity
What was there was using terminology not found anywhere else in the
spec (specifically “data profile”) and really didn’t make a great deal
of sense to me. But I’m pretty sure I understood what was intended, so I
completely rewrote it to something that makes sense.

I also added a bit indicating that urn:ietf:params:jmap:core shouldn’t
be present in primaryAccounts; it doesn’t makes much sense for it to be,
when you want there to potentially be different primary accounts for
different capabilities, and thus no one primary account.