Define primary accounts more carefully
This entails a few functional changes:

• Specification of which primary account to infer for null accountId;
• Returning accountNotFound when there is no suitable primary account;
• Email/copy, similar but for fromAccountId and toAccountId;
• Blob/copy, fromAccountId and toAccountId are made mandatory.

Details (including what and why) are in the narrative that follows.

I believe there used to be but one (exactly one?) primary account, but
now there is up to one per capability URI. Most of the spec hadn’t been
updated to match this reality, and there were a few places where it
caused problems of mild and easily resolved ambiguity, a couple of
places where it led to undefined or unimplementable behaviour.

Mostly this is just clarifying what a normal person would have assumed,
but it also defines something that was previously undefined: what to do
when no accountId (or fromAccountId or toAccountId for Email/copy and
Blob/copy) is passed, *and* no primary account is found: return
accountNotFound (or fromAccountNotFound or toAccountNotFound for
Email/copy and Blob/copy).

One other substantial change: Blob/copy’s fromAccountId and toAccountId
are now mandatory, because there’s no specification to tie Blob to (it’s
not a regular data type, but is fundamental to JMAP and used from other
specifications), so it can only make any sense while there is only one
primary account in the session. We *could* keep it optional, telling it
“if all the primary accounts are the same, breathe a sigh of relief and
use that primary account, otherwise panic”, but that would lead to
surprising behaviour when you add a new capability on the server and it
suddenly breaks something that formerly worked; that’s a bad mode of
extensibility, so it’s best to avoid the pain by making those particular
arguments mandatory from the start. Again, this is only going to be a
problem for Blob because it’s core; Email/copy doesn’t have this problem
because it’s tied to the urn:ietf:params:jmap:mail capability, which can
have a primary account.

I still confess myself unconvinced that accountId being optional is
desirable.

There’s inconsistency in the spec concerning what it calls things like
“urn:ietf:params:jmap:mail”; some places refer to it as a “specification
[URI]”, others as a “capability [URI]”. We should normalise all of these
at some point. I went for referring to them as capabilities in these
changes as I got the vague impression that was probably preferred.