Remove attachments-only search for now.
(The text search may still search inside attachments if the server supports this). Search inside attachments only is not easily implementable on top of existing IMAP standards, and is optional to support in the spec anyway so should be an extension instead. Servers could then advertise whether they support it or not.

Resolves https://github.com/jmapio/jmap/issues/221