Add GroupedAddresses parsed header form.
Reasoning as per Ricardo Signes:

1.  I love that JMAP makes working with email so approachable for non-domain
    experts.  It's an easy to understand API, and you can get to work by using
    it in a beginner's way before progressing, if ever, to more complex modes
    of operation.  I want to build on this property of JMAP when possible.

2.  I love that even for advanced operation, it is almost never necessary to
    resort to fetching the message/rfc822 blob.  Most data is available in raw
    or differently-processed forms.  Even experts shouldn't have to resort to
    blob processing most of the time.

Mailbox groups are worth making available because of reason #2.  Experts
shouldn't have to get the raw header, encoded-warts and all, and work out the
header structure.

On the other hand, mailbox groups are rarely seen and even more rarely worth
noting.  The casual user can, I assert, pay them no mind.  Groups should not
complicate the life of the casual user.

Providing a flat list is nice for the casual user, but if the flat list
contains weird-o non-address sentinel values, *it's no longer a flat list*.
It's a deeper structure that has been serialized into a flat list.  I think
it's worse for both sets of users: it's an unnatural expression of the
structured data that an advanced user might want, and it's an unnatural
expression of the simple data that a beginner might want.

(I also think that in the simple case, any entry with a null email is not
great.)

This is why I think we're best off providing two distinct forms:  one to
provide the simplest form for 99.9% of input and for casual users, and one with
an appropriate data structure for those who care about group structure.