Make Mailbox#unreadThreads property server-specific
There's at least two reasonable ways you could calculate this, the "easy"
way (which Gmail currently does) and the "right" way (which matches the
number of unread threads the user sees if they open the mailbox, and
FastMail currently does).

It seems most likely that the spec will be ignored if one way is
mandated, so instead make it server-dependent with a recommendation,
and specify clients are only to use it as an indication.