Make accountId arguments compulsory
Avoids potential issues if primary account changes at some point, and it's
easier for server and proxy implementors to deal with. The client must have
the account information if it has the API URL since they're both on the session
object, so it's easy for it to always explicitly include an id.