Add new workflowStatus and workflowStatuses
Before this I tried to extend progress with a bunch of common
values. This is probably not a good idea, because:
* the semantics of common values are system-dependent (e.g. resolved vs.
  completed or confirmed vs. in-process)
* there are event systems that even allow defining custom
  workflows on a project-level

I therefore introduce workflowStatuses on a folder-level to support
Jira's project-level workflow definition to some extent. From an
interoperability point-of-view this, however, is non-optimal as it will
make mapping workflowStatus difficult.